+++ package.json
@@ -85,8 +85,8 @@
       "^.+\\.(ts|tsx)$": "ts-jest"
     },
     "moduleNameMapper": {
-      "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/src/utility/assetsTransformer.js",
-      "\\.(css|less)$": "<rootDir>/src/utility/assetsTransformer.js"
+      "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/src/utils/assetsTransformer.js",
+      "\\.(css|less)$": "<rootDir>/src/utils/assetsTransformer.js"
     },
     "coverageThreshold": {
       "global": {
