+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Switch } from "react-router-dom"
 import { BrowserRouter as Router } from "react-router-dom"
-import { getRoutes } from "./utility/routeHelper"
+import { getRoutes } from "./utils/routeHelper"
 
 function App() {
   return (
